--- a/.prettierignore
+++ b/.prettierignore
@@ -1,2 +1,0 @@
-node_modules
-_site