--- a/.prettierrc
+++ b/.prettierrc
@@ -1,4 +1,4 @@
 {
-  "printWidth": 100,
-  "singleQuote": true
+    "printWidth": 100,
+    "singleQuote": true
 }