--- a/package.json
+++ b/package.json
@@ -3,77 +3,33 @@
 	"version": "0.1.0",
 	"description": "Bits und Bäume Website",
 	"scripts": {
-		"build-dev:webpack": "cross-env NODE_ENV=development webpack --config webpack.dev.js",
-		"watch:webpack": "webpack --watch --config webpack.dev.js",
-		"watch:eleventy": "cross-env ELEVENTY_ENV=development eleventy --serve",
-		"dev": "npm-run-all clean build-dev:webpack --parallel watch:*",
-		"build:webpack": "cross-env NODE_ENV=production webpack --config webpack.prod.js",
-		"build:eleventy": "cross-env ELEVENTY_ENV=production eleventy",
-		"build": "run-s clean build:* clean:thumbnails",
-		"serve:local": "serve _site",
-		"serve": "run-s build serve:local",
-		"clean": "rimraf _site",
-		"clean:thumbnails": "rimraf src/assets/images/gt--*.*",
-		"format:js": "prettier '**/*.js' --write && eslint '**/*.js' --fix",
-		"format:json": "prettier '**/*.json' --write",
-		"format": "run-s format:*",
-		"lint": "eslint --ext .js .",
-		"prepare": "husky install"
+		"dev:tailwind": "tailwindcss -i ./src/assets/styles/main.css -o ./src/assets/styles/tailwind.css -w",
+		"dev:eleventy": "eleventy --serve",
+		"dev": "run-p dev:*",
+		"build:tailwind": "tailwindcss -i ./src/assets/styles/main.css -o ./src/assets/styles/tailwind.css -m",
+		"build:eleventy": "eleventy",
+		"build": "npm-run-all build:*"
 	},
 	"license": "ISC",
 	"devDependencies": {
-		"@11ty/eleventy": "^1.0",
-		"autoprefixer": "^10.3.4",
-		"broken-link-checker": "^0.7.8",
-		"copy-webpack-plugin": "^8.1.1",
-		"cross-env": "^7.0.3",
-		"css-loader": "^5.2.7",
-		"cssnano": "^5.0.8",
-		"date-fns": "^2.24.0",
+		"@11ty/eleventy": "^1.0.2",
+		"autoprefixer": "^10.4.12",
+		"date-fns": "^2.29.3",
 		"eleventy-plugin-external-links": "^1.1.2",
 		"eleventy-plugin-img2picture": "^3.0.0",
-		"eslint": "^7.32.0",
-		"eslint-config-airbnb-base": "^14.2.1",
-		"eslint-plugin-import": "^2.24.2",
-		"file-loader": "^6.2.0",
 		"glob": "^7.1.7",
-		"hint": "^6.1.7",
 		"html-minifier": "^4.0.0",
-		"husky": "^6.0.0",
-		"image-webpack-loader": "^7.0.1",
-		"lint-staged": "^10.5.4",
-		"markdown-it-attrs": "^4.1.3",
-		"netlify-cms-app": "^2.14.36",
+		"markdown-it-attrs": "^4.1.4",
 		"npm-run-all": "^4.1.5",
-		"postcss-loader": "^5.3.0",
-		"prettier": "^2.4.1",
-		"rimraf": "^3.0.2",
-		"serve": "^11.3.2",
-		"tailwindcss": "^2.2.15",
-		"wait-for-localhost-cli": "^3.0.0",
-		"webpack-cli": "^4.8.0",
-		"webpack-merge": "^5.8.0"
-	},
-	"lint-staged": {
-		"*.js": [
-			"prettier --write",
-			"eslint --fix",
-			"eslint"
-		],
-		"*.json": [
-			"prettier --write"
-		]
+		"postcss": "^8.4.18",
+		"tailwindcss": "^3.2.1"
 	},
 	"dependencies": {
 		"debug": "^4.3.4",
-		"glob": "^7.1.7",
-		"html-webpack-plugin": "^5.3.2",
+		"glob": "^7.2.3",
 		"js-yaml": "^4.1.0",
-		"mini-css-extract-plugin": "^1.6.2",
-		"npm": "^7.24.0",
-		"sharp": "^0.30.7",
-		"webpack": "^5.53.0",
-		"webpack-remove-empty-scripts": "^0.7.1"
+		"markdown-it": "^13.0.1",
+		"sharp": "^0.30.7"
 	},
 	"browserslist": [
 		"> 2% in DE"