--- a/postcss.config.js
+++ b/postcss.config.js
@@ -1,19 +1,6 @@
-/* eslint-disable import/no-extraneous-dependencies, global-require */
-const plugins = [require('tailwindcss'), require('autoprefixer')];
-
-if (process.env.NODE_ENV === 'production') {
-  plugins.push(
-    require('cssnano')({
-      preset: [
-        'default',
-        {
-          discardComments: {
-            removeAll: true,
-          },
-        },
-      ],
-    }),
-  );
-}
-
-module.exports = { plugins };
+module.exports = {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+};