--- a/webpack.html
+++ b/webpack.html
@@ -1,4 +1,0 @@
-<%= htmlWebpackPlugin.files.css.map((css) => `<link
-  href="/assets${css}"
-  rel="stylesheet"
-/>`).join('') %>