# Issue #47: Build process is slow and getting slower

## Removed unwanted node modules

Removed the following node modules since they serve no real purpose in the project.

```json
"broken-link-checker": "^0.7.8",
"copy-webpack-plugin": "^8.1.1",
"cross-env": "^7.0.3",
"css-loader": "^5.2.7",
"cssnano": "^5.0.8",
"eslint": "^7.32.0",
"eslint-config-airbnb-base": "^14.2.1",
"eslint-plugin-import": "^2.24.2",
"file-loader": "^6.2.0",
"hint": "^6.1.7",
"html-webpack-plugin": "^5.3.2",
"husky": "^6.0.0",
"image-webpack-loader": "^7.0.1",
"lint-staged": "^10.5.4",
"markdown-it-attrs": "^4.1.3",
"mini-css-extract-plugin": "^1.6.2",
"netlify-cms-app": "^2.14.36",
"npm": "^7.24.0",
"postcss-loader": "^5.3.0",
"prettier": "^2.4.1",
"rimraf": "^3.0.2",
"serve": "^11.3.2",
"wait-for-localhost-cli": "^3.0.0",
"webpack": "^5.53.0",
"webpack-cli": "^4.8.0",
"webpack-merge": "^5.8.0"
"webpack-remove-empty-scripts": "^0.7.1"
```

## Updated npm scripts

Replaced unwanted scripts with only useful, performant ones.

### Before

```json
"build-dev:webpack": "cross-env NODE_ENV=development webpack --config webpack.dev.js",
"watch:webpack": "webpack --watch --config webpack.dev.js",
"watch:eleventy": "cross-env ELEVENTY_ENV=development eleventy --serve",
"dev": "npm-run-all clean build-dev:webpack --parallel watch:*",
"build:webpack": "cross-env NODE_ENV=production webpack --config webpack.prod.js",
"build:eleventy": "cross-env ELEVENTY_ENV=production eleventy",
"build": "run-s clean build:* clean:thumbnails",
"serve:local": "serve _site",
"serve": "run-s build serve:local",
"clean": "rimraf _site",
"clean:thumbnails": "rimraf src/assets/images/gt--*.*",
"format:js": "prettier '**/*.js' --write && eslint '**/*.js' --fix",
"format:json": "prettier '**/*.json' --write",
"format": "run-s format:*",
"lint": "eslint --ext .js .",
"prepare": "husky install"
```

### After

```json
"dev:tailwind": "tailwindcss -i ./src/assets/styles/main.css -o ./src/assets/styles/tailwind.css -w",
"dev:eleventy": "eleventy --serve",
"dev": "concurrently \"npm:dev:*\"",
"build:tailwind": "tailwindcss -i ./src/assets/styles/main.css -o ./src/assets/styles/tailwind.css -m",
"build:eleventy": "eleventy",
"build": "npm-run-all build:*"
```

## Reduced build time to ~17 seconds

### Before

```
$ export DEBUG="Eleventy:Benchmark* npx @11ty/eleventy"
$ npm run dev

> bits-und-baeume@0.1.0 dev
> npm-run-all clean build-dev:webpack --parallel watch:*


> bits-und-baeume@0.1.0 clean
> rimraf _site


> bits-und-baeume@0.1.0 build-dev:webpack
> cross-env NODE_ENV=development webpack --config webpack.dev.js

asset styles/main.css 1.71 MiB [emitted] (name: main)
asset images/9d9b77f868ec6f4f6204.svg 857 bytes [emitted] [immutable] [from: src/assets/images/up-arrow.svg] (auxiliary name: main)
asset ../../src/_includes/partials/_webpack.ejs 61 bytes [compared for emit]
runtime modules 274 bytes 1 module
./src/assets/styles/main.css 50 bytes [built] [code generated]
css ./node_modules/css-loader/dist/cjs.js!./node_modules/postcss-loader/dist/cjs.js!./src/assets/styles/main.css 1.71 MiB [built] [code generated]
webpack 5.73.0 compiled successfully in 8964 ms

> bits-und-baeume@0.1.0 watch:eleventy
> cross-env ELEVENTY_ENV=development eleventy --serve


> bits-und-baeume@0.1.0 watch:webpack
> webpack --watch --config webpack.dev.js

assets by status 857 bytes [cached] 1 asset
asset styles/main.css 1.71 MiB [compared for emit] (name: main)
asset ../../src/_includes/partials/_webpack.ejs 61 bytes [compared for emit]
runtime modules 274 bytes 1 module
./src/assets/styles/main.css 50 bytes [built] [code generated]
css ./node_modules/css-loader/dist/cjs.js!./node_modules/postcss-loader/dist/cjs.js!./src/assets/styles/main.css 1.71 MiB [built] [code generated]
webpack 5.73.0 compiled successfully in 28190 ms
  Eleventy:Benchmark Benchmark      1ms   0%    18× (Configuration) "slug" Nunjucks Filter +0ms
  Eleventy:Benchmark Benchmark      1ms   0%    40× (Configuration) "basefilename" Nunjucks Filter +2ms
  Eleventy:Benchmark Benchmark     18ms   0%    33× (Configuration) "thumbnail" Nunjucks Async Shortcode +0ms
  Eleventy:Benchmark Benchmark     59ms   0%     2× (Aggregate) Searching the file system +0ms
  Eleventy:Benchmark Benchmark     13ms   0%    29× (Aggregate) Data File +0ms
  Eleventy:Benchmark Benchmark    361ms   1%   142× (Aggregate) Template Read +0ms
  Eleventy:Benchmark Benchmark    266ms   0%   152× (Aggregate) Template Compile +0ms
  Eleventy:Benchmark Benchmark     12ms   0%     3× (Aggregate) > Compile > ./src/gallery-main-images.njk +0ms
  Eleventy:Benchmark Benchmark  49223ms  80%   129× (Aggregate) Render +0ms
  Eleventy:Benchmark Benchmark      5ms   0%    30× (Aggregate) > Render > ./src/gallery-main-images.njk (Pagination: 6 pages) +0ms
  Eleventy:Benchmark Benchmark  36345ms  59%   214× (Aggregate) Passthrough Copy File +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > _site/gallery/main/gallery-example-1/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Compile > ./src/konferenz-2022/awareness.njk +1ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > ./src/konferenz-2022/gallerie-presse.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > ./src/konferenz-2022/gallerie.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > ./src/konferenz-2022/material.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > ./src/konferenz-2022/ueber-uns.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Compile > ./src/en/conference-2022/awareness.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > ./src/en/conference-2022/about-us.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     2× (Aggregate) > Render > ./src/en/conference-2022/demands.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     2× (Aggregate) > Render > ./src/konferenz-2022/forderungen.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     2× (Aggregate) > Render > ./src/konferenz-2022/faqs.njk +0ms
  Eleventy:Benchmark Benchmark      4ms   0%     2× (Aggregate) > Compile > ./src/404.md +0ms
  Eleventy:Benchmark Benchmark      7ms   0%     2× (Aggregate) > Render > ./src/404.md +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     2× (Aggregate) > Compile > ./src/robots.ejs +0ms
  Eleventy:Benchmark Benchmark      6ms   0%     1× (Aggregate) > Compile > ./src/__konferenz-2022-old.njk +0ms
  Eleventy:Benchmark Benchmark     30ms   0%     1× (Aggregate) > Compile > ./src/index.njk +0ms
  Eleventy:Benchmark Benchmark  19015ms  31%     1× (Aggregate) > Render > ./src/index.njk +0ms
  Eleventy:Benchmark Benchmark  19015ms  31%     1× (Aggregate) > Render > _site/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/konferenz-2022.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/index.html +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/impressum.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/klimabewusste-website.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/lost-and-found.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/community/bildung.njk +0ms
  Eleventy:Benchmark Benchmark      9ms   0%     1× (Aggregate) > Compile > ./src/community/eigene-organisieren.njk +1ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Render > ./src/community/eigene-organisieren.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Render > _site/community/eigene-organisieren/index.html +0ms
  Eleventy:Benchmark Benchmark     11ms   0%     1× (Aggregate) > Compile > ./src/community/forum.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/community/forum.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/community/forum/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/community/index.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/community/regionale-zweige.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/community/regionale-zweige.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/community/regionale-zweige/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/community/konferenz-nrw.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/forderungen.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Render > ./src/konferenz-2018/forderungen.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Render > _site/konferenz-2018/forderungen/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/index.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/nachhaltige-konferenzorga.njk +0ms
  Eleventy:Benchmark Benchmark     10ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/presse.njk +0ms
  Eleventy:Benchmark Benchmark   8249ms  13%     1× (Aggregate) > Render > ./src/konferenz-2018/presse.njk +0ms
  Eleventy:Benchmark Benchmark   8249ms  13%     1× (Aggregate) > Render > _site/konferenz-2018/presse/index.html +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/programm.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/konferenz-2018/programm.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2018/programm/index.html +0ms
  Eleventy:Benchmark Benchmark      5ms   0%     1× (Aggregate) > Compile > ./src/en/index.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/en/index.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/ziele.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/en/conference-2022.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/conference-2022/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/__conference-2022-old.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/lost-and-found.njk +0ms
  Eleventy:Benchmark Benchmark      9ms   0%     1× (Aggregate) > Compile > ./src/elemente.njk +0ms
  Eleventy:Benchmark Benchmark  19691ms  32%     1× (Aggregate) > Render > ./src/elemente.njk +1ms
  Eleventy:Benchmark Benchmark  19691ms  32%     1× (Aggregate) > Render > _site/elemente/index.html +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2018/material/index.njk +0ms
  Eleventy:Benchmark Benchmark      9ms   0%     1× (Aggregate) > Compile > ./src/en/community/forum.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/en/community/forum.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/community/forum/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/community/index.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/community/konferenz-nrw.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/en/community/regional-branches.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/en/community/regional-branches.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/community/regional-branches/index.html +0ms
  Eleventy:Benchmark Benchmark      4ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2018/demands.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/en/conference-2018/demands.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/conference-2018/demands/index.html +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/en/community/self-organisation.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2018/goals.njk +0ms
  Eleventy:Benchmark Benchmark      5ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2018/press.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/en/conference-2018/press.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/conference-2018/press/index.html +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2018/program.njk +1ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2018/sustainable-orga.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/material/index.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/gallerie-presse.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/gallerie-presse/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/gallerie/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/karte.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/mach-mit.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/material.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/material/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/nachhaltige-konferenzorga.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/presse_alt.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/presse.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/programm.njk +1ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/tickets.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/ueber-uns.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/ueber-uns/index.html +0ms
  Eleventy:Benchmark Benchmark      4ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/about-us.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/en/conference-2022/about-us/index.html +0ms
  Eleventy:Benchmark Benchmark      6ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/demands.njk +0ms
  Eleventy:Benchmark Benchmark      2ms   0%     1× (Aggregate) > Render > _site/en/conference-2022/demands/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/faqs.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/map.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/press.njk +0ms
  Eleventy:Benchmark Benchmark      4ms   0%     1× (Aggregate) > Compile > ./src/datenschutz.md +0ms
  Eleventy:Benchmark Benchmark      8ms   0%     1× (Aggregate) > Render > ./src/datenschutz.md +0ms
  Eleventy:Benchmark Benchmark      8ms   0%     1× (Aggregate) > Render > _site/datenschutz/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/program.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/en/conference-2022/tickets.njk +0ms
  Eleventy:Benchmark Benchmark     10ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/forderungen.njk +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/forderungen/index.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/konferenz-2022/faqs.njk +1ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/konferenz-2022/faqs/index.html +0ms
  Eleventy:Benchmark Benchmark      6ms   0%     1× (Aggregate) > Render > _site/404.html +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Compile > ./src/sitemap.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > ./src/sitemap.njk +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Aggregate) > Render > _site/sitemap.xml +0ms
  Eleventy:Benchmark Benchmark    810ms   1%     1× (Aggregate) > Render > ./src/en/menu.ejs +0ms
  Eleventy:Benchmark Benchmark    810ms   1%     1× (Aggregate) > Render > _site/en/menu/index.html +0ms
  Eleventy:Benchmark Benchmark   1404ms   2%     1× (Aggregate) > Render > ./src/menu.ejs +0ms
  Eleventy:Benchmark Benchmark   1404ms   2%     1× (Aggregate) > Render > _site/menu/index.html +0ms
  Eleventy:Benchmark Benchmark   8524ms  14%    76× (Aggregate) Template Write +0ms
  Eleventy:Benchmark Benchmark     57ms   0%    72× (Aggregate) > Compile > ./src/_includes/layouts/_base.ejs +0ms
  Eleventy:Benchmark Benchmark     38ms   0%     4× (Aggregate) > Compile > ./src/_includes/layouts/_page.ejs +0ms
  Eleventy:Benchmark Benchmark      3ms   0%     2× (Aggregate) > Compile > ./src/_includes/layouts/_mobile-menu.ejs +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Data) `./src/_data/site.json` +0ms
  Eleventy:Benchmark Benchmark      1ms   0%     1× (Data) `./src/_data/layout.js` +0ms
  Eleventy:Benchmark Benchmark      5ms   0%    16× (Data) `./src/konferenz-2022/konferenz-2022.json` +1ms
  Eleventy:Benchmark Benchmark      4ms   0%    11× (Data) `./src/en/conference-2022/conference-2022.json` +0ms
[11ty] Copied 214 files / Wrote 76 files in 58.87 seconds (774.6ms each, v1.0.1)
  Eleventy:Benchmark Benchmark   1478ms NaN%     1× (Watch) Start up --watch +1s
  Eleventy:Benchmark Benchmark   1461ms NaN%     1× (Watch) Watching JavaScript Dependencies (disable with `eleventyConfig.setWatchJavaScriptDependencies(false)`) +0ms
[11ty] Watching…
[Browsersync] Access URLs:
 -------------------------------------
    Local: http://localhost:8080
 External: http://192.168.209.243:8080
 -------------------------------------
[Browsersync] Serving files from: _site
[Browsersync] Watching files...
```

### After

```
> bits-und-baeume@0.1.0 build
> npm-run-all build:*


> bits-und-baeume@0.1.0 build:tailwind
> tailwindcss -i ./src/assets/styles/main.css -o ./src/assets/styles/tailwind.css -m


Rebuilding...

Done in 1506ms.

> bits-und-baeume@0.1.0 build:eleventy
> eleventy

[11ty] Writing _site/robots.txt from ./src/robots.ejs
[11ty] Writing _site/sitemap.xml from ./src/sitemap.njk
…