@cknoll 2022-10-25
For those who want to try out this PR I used the following commands:

`git fetch origin pull/49/head:pr49`
`git checkout pr49`

While i can confirm that the build process is faster, it is (mostly) still surprisingly slow on my machine. However the results are inconsistent: I get times between 14s (one time) and 70s (one time) with an average of ca. 40s.

I see two open questions:
- a) Does this PR unintendedly remove something important? -> @michael-voit should be able to answer this
- b) why is the built process still so slow and not even consistently?

I suggest to focus discussion here on a) and discuss b) in the corresponding issuse #47.