@michael-voit 2022-10-25
@cknoll It is a hard question, if this PR removes something important.... That depends. :) What we can try to see is, what will be changed by it (a question the developer has done the changes should be best suited to answer).

An a first quick view into it I would say:
- Some reductions are not in any relation to build time in my understanding (without testing it) - but taking away tools and possibilites. For example:
  - remove of several config files and linters: code can no longer conform easily to a common standard and will not be linted before commits - both can reduce code quality with time.
  - Removal of npm packages like "hint" or "broken link checker" prevents the website from being checked quickly and in the working environment for the most significant errors before a push into the main branch.
- netlifyCMS removed - was not used anyway
- PostCSS changed - looks like there could be less optimization resulting in bigger CSS file - should be tested
- rimraf removed - that means, that the build folder is not deleted before a new build. Old files that no longer exist in new version can still exist in build folder than, so that it is not a clean build on local machine - and so errors after fresh builds could remain undetected in some circumstances.
- changes in npm scripts, so the build process has to be changed accordingly
- removal of webpack takes away a lot of possibilities

All in all I would say that this is a great reduction of stuff - not all related to build time. If all of them are "imortant" or which of them should better stay depends on meanings and ideas.

I would never merge such a big change directly into the main branch. I would create a new branch to merge that in - and than A/B test your way of working and the resulting file sizes and so on between the site it is now and the site resulting after this changes. You have to see if you would have testing capabilities, quality assurance baked in and so on. Could be, that this PR is exactly right and great for you. For me, it wouldn't be.

Please don't see just build time as the only measure. Much more important is quality and size of the result. And that has to be tested.